+++ lib/cgo/coin.block.go
@@ -342,3 +342,70 @@ func SKY_coin_NewEmptyBlock(_txns C.Transactions__Handle, handle *C.Block__Handl
 	*handle = registerBlockHandle(&block)
 	return
 }
+
+//export SKY_coin_Block_GetBlockHeader
+func SKY_coin_Block_GetBlockHeader(_b C.Block__Handle, _bh *C.BlockHeader__Handle) (____error_code uint32) {
+	b, ok := lookupBlockHandle(_b)
+	if !ok {
+		____error_code = SKY_BAD_HANDLE
+	} else {
+		*_bh = registerBlockHeaderHandle(&b.Head)
+	}
+	return
+}
+
+//export SKY_coin_GetBlockHeaderObject
+func SKY_coin_GetBlockHeaderObject(_b C.BlockHeader__Handle, _p **C.coin__BlockHeader) (____error_code uint32) {
+	b, ok := lookupBlockHeaderHandle(_b)
+	if !ok {
+		____error_code = SKY_BAD_HANDLE
+	} else {
+		*_p = (*C.coin__BlockHeader)(unsafe.Pointer(b))
+	}
+	return
+}
+
+//export SKY_coin_BlockHeader_Time
+func SKY_coin_BlockHeader_Time(_b C.BlockHeader__Handle, _arg0 *uint64) (____error_code uint32) {
+	b, ok := lookupBlockHeaderHandle(_b)
+	if !ok {
+		____error_code = SKY_BAD_HANDLE
+	} else {
+		*_arg0 = uint64(b.Time)
+	}
+	return
+}
+
+//export SKY_coin_BlockHeader_BkSeq
+func SKY_coin_BlockHeader_BkSeq(_b C.BlockHeader__Handle, _arg0 *uint64) (____error_code uint32) {
+	b, ok := lookupBlockHeaderHandle(_b)
+	if !ok {
+		____error_code = SKY_BAD_HANDLE
+	} else {
+		*_arg0 = uint64(b.BkSeq)
+	}
+	return
+}
+
+//export SKY_coin_BlockHeader_UxHash
+func SKY_coin_BlockHeader_UxHash(_b C.BlockHeader__Handle, _arg0 *C.cipher__SHA256) (____error_code uint32) {
+	b, ok := lookupBlockHeaderHandle(_b)
+	if !ok {
+		____error_code = SKY_BAD_HANDLE
+	} else {
+		*_arg0 = *(*C.cipher__SHA256)(unsafe.Pointer(&b.UxHash))
+	}
+	return
+}
+
+//export SKY_coin_BlockBody_Transactions
+func SKY_coin_BlockBody_Transactions(_bb C.BlockBody__Handle, _arg0 *C.Transactions__Handle) (____error_code uint32) {
+	bb, ok := lookupBlockBodyHandle(_bb)
+	if !ok {
+		____error_code = SKY_BAD_HANDLE
+		return
+	}
+	__arg0 := bb.Transactions
+	*_arg0 = registerTransactionsHandle(&__arg0)
+	return
+}