+++ lib/cgo/libsky_handle.go
@@ -383,3 +383,17 @@ func lookupCreateTransactionParamsHandle(handle C.CreateTransactionParams__Handl
 	}
 	return nil, false
 }
+
+func registerBlockHeaderHandle(obj *coin.BlockHeader) C.BlockHeader__Handle {
+	return (C.BlockHeader__Handle)(registerHandle(obj))
+}
+
+func lookupBlockHeaderHandle(handle C.BlockHeader__Handle) (*coin.BlockHeader, bool) {
+	obj, ok := lookupHandle(C.Handle(handle))
+	if ok {
+		if obj, isOK := (obj).(*coin.BlockHeader); isOK {
+			return obj, true
+		}
+	}
+	return nil, false
+}