+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Add function `SKY_coin_BlockBody_Transactions`
+- Add function `SKY_coin_BlockHeader_UxHash`
+- Add function `SKY_coin_BlockHeader_BkSeq`
+- Add function `SKY_coin_BlockHeader_Time`
+- Add function `SKY_coin_GetBlockHeaderObject`
+- Add function `SKY_coin_Block_GetBlockHeader`
 - Add function `SKY_base58_Hex2Base58`
 - Add function `SKY_base58_Encode`
 - Add function `SKY_base58_Decode`